requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
pysrt==1.1.2
google-cloud-speech==2.21.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2